Crate rustc_trait_selection

This crate defines the trait resolution method.

Traits. Trait resolution is implemented in the traits module.

For more information about how rustc works, see the rustc-dev-guide.

Note

This API is completely unstable and subject to change.

Modules

error_reporting

errors

infer

regions

solve

traits

Trait Resolution. See the rustc dev guide for more information on how this works.

Statics

DEFAULT_LOCALE_RESOURCE

Raw content of Fluent resource for this crate, generated by fluent_messages macro, imported by rustc_driver to include all crates’ resources in one bundle.