rustc_trait_selection::error_reporting::infer::nice_region_error
Module different_lifetimes

Error Reporting for Anonymous Region Lifetime Errors where both the regions are anonymous.

Functions

suggest_adding_lifetime_params

Currently only used in rustc_borrowck, probably should be removed in favour of public_errors::AddLifetimeParamsSuggestion