Struct rustc_trait_selection::error_reporting::TypeErrCtxt

pub struct TypeErrCtxt<'a, 'tcx> {
    pub infcx: &'a InferCtxt<'tcx>,
    pub sub_relations: RefCell<SubRelations>,
    pub typeck_results: Option<Ref<'a, TypeckResults<'tcx>>>,
    pub fallback_has_occurred: bool,
    pub normalize_fn_sig: Box<dyn Fn(PolyFnSig<'tcx>) -> PolyFnSig<'tcx> + 'a>,
    pub autoderef_steps: Box<dyn Fn(Ty<'tcx>) -> Vec<(Ty<'tcx>, Vec<PredicateObligation<'tcx>>)> + 'a>,
}

A helper for building type related errors. The typeck_results field is only populated during an in-progress typeck. Get an instance by calling InferCtxt::err_ctxt or FnCtxt::err_ctxt.

You must only create this if you intend to actually emit an error (or perhaps a warning, though preferably not.) It provides a lot of utility methods which should not be used during the happy path.

Fields
infcx: &'a InferCtxt<'tcx>sub_relations: RefCell<SubRelations>typeck_results: Option<Ref<'a, TypeckResults<'tcx>>>fallback_has_occurred: boolnormalize_fn_sig: Box<dyn Fn(PolyFnSig<'tcx>) -> PolyFnSig<'tcx> + 'a>autoderef_steps: Box<dyn Fn(Ty<'tcx>) -> Vec<(Ty<'tcx>, Vec<PredicateObligation<'tcx>>)> + 'a>
Implementations

impl<'tcx> TypeErrCtxt<'_, 'tcx>

pub fn note_and_explain_type_err( &self, diag: &mut Diag<'_>, err: TypeError<'tcx>, cause: &ObligationCause<'tcx>, sp: Span, body_owner_def_id: DefId, )

fn suggest_constraint( &self, diag: &mut Diag<'_>, msg: impl Fn() -> String, body_owner_def_id: DefId, proj_ty: AliasTy<'tcx>, ty: Ty<'tcx>, ) -> bool

fn expected_projection( &self, diag: &mut Diag<'_>, proj_ty: AliasTy<'tcx>, values: ExpectedFound<Ty<'tcx>>, body_owner_def_id: DefId, cause_code: &ObligationCauseCode<'_>, )

An associated type was expected and a different type was found.

We perform a few different checks to see what we can suggest:

In the current item, look for associated functions that return the expected type and suggest calling them. (Not a structured suggestion.)

If any of the item’s generic bounds can be constrained, we suggest constraining the associated type to the found type.

If the associated type has a default type and was expected inside of a trait, we mention that this is disallowed.

If all other things fail, and the error is not because of a mismatch between the trait and the impl, we provide a generic help to constrain the assoc type or call an assoc fn that returns the type.

fn suggest_constraining_opaque_associated_type( &self, diag: &mut Diag<'_>, msg: impl Fn() -> String, proj_ty: AliasTy<'tcx>, ty: Ty<'tcx>, ) -> bool

When the expected impl Trait is not defined in the current item, it will come from a return type. This can occur when dealing with TryStream (#71035).

fn point_at_methods_that_satisfy_associated_type( &self, diag: &mut Diag<'_>, assoc_container_id: DefId, current_method_ident: Option<Symbol>, proj_ty_item_def_id: DefId, expected: Ty<'tcx>, ) -> bool

fn point_at_associated_type( &self, diag: &mut Diag<'_>, body_owner_def_id: DefId, found: Ty<'tcx>, ) -> bool

fn constrain_generic_bound_associated_type_structured_suggestion( &self, diag: &mut Diag<'_>, trait_ref: TraitRef<'tcx>, bounds: GenericBounds<'_>, assoc: AssocItem, assoc_args: &[GenericArg<'tcx>], ty: Ty<'tcx>, msg: impl Fn() -> String, is_bound_surely_present: bool, ) -> bool

Given a slice of hir::GenericBounds, if any of them corresponds to the trait_ref requirement, provide a structured suggestion to constrain it to a given type ty.

is_bound_surely_present indicates whether we know the bound we’re looking for is inside bounds. If that’s the case then we can consider bounds containing only one trait bound as the one we’re looking for. This can help in cases where the associated type is defined on a supertrait of the one present in the bounds.

fn constrain_associated_type_structured_suggestion( &self, diag: &mut Diag<'_>, span: Span, assoc: AssocItem, assoc_args: &[GenericArg<'tcx>], ty: Ty<'tcx>, msg: impl Fn() -> String, ) -> bool

Given a span corresponding to a bound, provide a structured suggestion to set an associated type to a given type ty.

pub fn format_generic_args(&self, args: &[GenericArg<'tcx>]) -> String

impl<'tcx> TypeErrCtxt<'_, 'tcx>

pub(super) fn suggest_remove_semi_or_return_binding( &self, first_id: Option<HirId>, first_ty: Ty<'tcx>, first_span: Span, second_id: Option<HirId>, second_ty: Ty<'tcx>, second_span: Span, ) -> Option<SuggestRemoveSemiOrReturnBinding>

pub(super) fn suggest_tuple_pattern( &self, cause: &ObligationCause<'tcx>, exp_found: &ExpectedFound<Ty<'tcx>>, diag: &mut Diag<'_>, )

pub(super) fn suggest_await_on_expect_found( &self, cause: &ObligationCause<'tcx>, exp_span: Span, exp_found: &ExpectedFound<Ty<'tcx>>, diag: &mut Diag<'_>, )

A possible error is to forget to add .await when using futures:

async fn make_u32() -> u32 {
    22
}

fn take_u32(x: u32) {}

async fn foo() {
    let x = make_u32();
    take_u32(x);
}

This routine checks if the found type T implements Future<Output=U> where U is the expected type. If this is the case, and we are inside of an async body, it suggests adding .await to the tail of the expression.

pub(super) fn suggest_accessing_field_where_appropriate( &self, cause: &ObligationCause<'tcx>, exp_found: &ExpectedFound<Ty<'tcx>>, diag: &mut Diag<'_>, )

pub(super) fn suggest_turning_stmt_into_expr( &self, cause: &ObligationCause<'tcx>, exp_found: &ExpectedFound<Ty<'tcx>>, diag: &mut Diag<'_>, )

pub fn consider_removing_semicolon( &self, blk: &'tcx Block<'tcx>, expected_ty: Ty<'tcx>, diag: &mut Diag<'_>, ) -> bool

A common error is to add an extra semicolon:

fn foo() -> usize {
    22;
}

This routine checks if the final statement in a block is an expression with an explicit semicolon whose type is compatible with expected_ty. If so, it suggests removing the semicolon.

pub(super) fn suggest_function_pointers( &self, cause: &ObligationCause<'tcx>, span: Span, exp_found: &ExpectedFound<Ty<'tcx>>, diag: &mut Diag<'_>, )

pub fn should_suggest_as_ref_kind( &self, expected: Ty<'tcx>, found: Ty<'tcx>, ) -> Option<SuggestAsRefKind>

pub fn should_suggest_as_ref( &self, expected: Ty<'tcx>, found: Ty<'tcx>, ) -> Option<&str>

pub(super) fn suggest_let_for_letchains( &self, cause: &ObligationCause<'_>, span: Span, ) -> Option<TypeErrorAdditionalDiags>

Try to find code with pattern if Some(..) = expr use a visitor to mark the if which its span contains given error span, and then try to find a assignment in the cond part, which span is equal with error span

pub(super) fn suggest_for_all_lifetime_closure( &self, span: Span, hir: Node<'_>, exp_found: &ExpectedFound<TraitRef<'tcx>>, diag: &mut Diag<'_>, )

For “one type is more general than the other” errors on closures, suggest changing the lifetime of the parameters to accept all lifetimes.

impl<'tcx> TypeErrCtxt<'_, 'tcx>

pub fn could_remove_semicolon( &self, blk: &'tcx Block<'tcx>, expected_ty: Ty<'tcx>, ) -> Option<(Span, StatementAsExpression)>

Be helpful when the user wrote {... expr; } and taking the ; off is enough to fix the error.

pub fn consider_returning_binding_diag( &self, blk: &'tcx Block<'tcx>, expected_ty: Ty<'tcx>, ) -> Option<SuggestRemoveSemiOrReturnBinding>

Suggest returning a local binding with a compatible type if the block has no return expression.

pub fn consider_returning_binding( &self, blk: &'tcx Block<'tcx>, expected_ty: Ty<'tcx>, err: &mut Diag<'_>, ) -> bool

impl<'a, 'tcx> TypeErrCtxt<'a, 'tcx>

pub fn extract_inference_diagnostics_data( &self, arg: GenericArg<'tcx>, highlight: Option<RegionHighlightMode<'tcx>>, ) -> InferenceDiagnosticsData

Extracts data used by diagnostic for either types or constants which were stuck during inference.

fn bad_inference_failure_err( &self, span: Span, arg_data: InferenceDiagnosticsData, error_code: TypeAnnotationNeeded, ) -> Diag<'a>

Used as a fallback in TypeErrCtxt::emit_inference_failure_err in case we weren’t able to get a better error.

pub fn emit_inference_failure_err( &self, body_def_id: LocalDefId, failure_span: Span, arg: GenericArg<'tcx>, error_code: TypeAnnotationNeeded, should_label_span: bool, ) -> Diag<'a>

impl<'cx, 'tcx> TypeErrCtxt<'cx, 'tcx>

pub fn try_report_nice_region_error( &'cx self, generic_param_scope: LocalDefId, error: &RegionResolutionError<'tcx>, ) -> Option<ErrorGuaranteed>

impl<'a, 'tcx> TypeErrCtxt<'a, 'tcx>

pub fn report_region_errors( &self, generic_param_scope: LocalDefId, errors: &[RegionResolutionError<'tcx>], ) -> ErrorGuaranteed

fn process_errors( &self, errors: &[RegionResolutionError<'tcx>], ) -> Vec<RegionResolutionError<'tcx>>

pub(super) fn note_region_origin( &self, err: &mut Diag<'_>, origin: &SubregionOrigin<'tcx>, )

pub(super) fn report_concrete_failure( &self, generic_param_scope: LocalDefId, origin: SubregionOrigin<'tcx>, sub: Region<'tcx>, sup: Region<'tcx>, ) -> Diag<'a>

pub fn suggest_copy_trait_method_bounds( &self, trait_item_def_id: DefId, impl_item_def_id: LocalDefId, err: &mut Diag<'_>, )

pub(super) fn report_placeholder_failure( &self, generic_param_scope: LocalDefId, placeholder_origin: SubregionOrigin<'tcx>, sub: Region<'tcx>, sup: Region<'tcx>, ) -> Diag<'a>

pub fn report_generic_bound_failure( &self, generic_param_scope: LocalDefId, span: Span, origin: Option<SubregionOrigin<'tcx>>, bound_kind: GenericKind<'tcx>, sub: Region<'tcx>, ) -> ErrorGuaranteed

pub fn construct_generic_bound_failure( &self, generic_param_scope: LocalDefId, span: Span, origin: Option<SubregionOrigin<'tcx>>, bound_kind: GenericKind<'tcx>, sub: Region<'tcx>, ) -> Diag<'a>

pub fn suggest_name_region( &self, generic_param_scope: LocalDefId, lifetime: Region<'tcx>, add_lt_suggs: &mut Vec<(Span, String)>, ) -> String

fn report_sub_sup_conflict( &self, generic_param_scope: LocalDefId, var_origin: RegionVariableOrigin, sub_origin: SubregionOrigin<'tcx>, sub_region: Region<'tcx>, sup_origin: SubregionOrigin<'tcx>, sup_region: Region<'tcx>, ) -> ErrorGuaranteed

fn report_inference_failure(&self, var_origin: RegionVariableOrigin) -> Diag<'_>

impl<'a, 'tcx> TypeErrCtxt<'a, 'tcx>

pub fn type_error_struct_with_diag<M>( &self, sp: Span, mk_diag: M, actual_ty: Ty<'tcx>, ) -> Diag<'a>
where M: FnOnce(String) -> Diag<'a>,

pub fn report_mismatched_types( &self, cause: &ObligationCause<'tcx>, expected: Ty<'tcx>, actual: Ty<'tcx>, err: TypeError<'tcx>, ) -> Diag<'a>

pub fn report_mismatched_consts( &self, cause: &ObligationCause<'tcx>, expected: Const<'tcx>, actual: Const<'tcx>, err: TypeError<'tcx>, ) -> Diag<'a>

pub fn get_impl_future_output_ty(&self, ty: Ty<'tcx>) -> Option<Ty<'tcx>>

fn check_and_note_conflicting_crates( &self, err: &mut Diag<'_>, terr: TypeError<'tcx>, )

Adds a note if the types come from similarly named crates

fn note_error_origin( &self, err: &mut Diag<'_>, cause: &ObligationCause<'tcx>, exp_found: Option<ExpectedFound<Ty<'tcx>>>, terr: TypeError<'tcx>, )

fn highlight_outer( &self, value: &mut DiagStyledString, other_value: &mut DiagStyledString, name: String, sub: GenericArgsRef<'tcx>, pos: usize, other_ty: Ty<'tcx>, )

Given that other_ty is the same as a type argument for name in sub, populate value highlighting name and every type argument that isn’t at pos (which is other_ty), and populate other_value with other_ty.

Foo<Bar<Qux>>
^^^^--------^ this is highlighted
|   |
|   this type argument is exactly the same as the other type, not highlighted
this is highlighted
Bar<Qux>
-------- this type is the same as a type argument in the other type, not highlighted

fn cmp_type_arg( &self, t1_out: &mut DiagStyledString, t2_out: &mut DiagStyledString, path: String, sub: &'tcx [GenericArg<'tcx>], other_path: String, other_ty: Ty<'tcx>, ) -> Option<()>

If other_ty is the same as a type argument present in sub, highlight path in t1_out, as that is the difference to the other type.

For the following code:

let x: Foo<Bar<Qux>> = foo::<Bar<Qux>>();

The type error output will behave in the following way:

Foo<Bar<Qux>>
^^^^--------^ this is highlighted
|   |
|   this type argument is exactly the same as the other type, not highlighted
this is highlighted
Bar<Qux>
-------- this type is the same as a type argument in the other type, not highlighted

fn push_comma( &self, value: &mut DiagStyledString, other_value: &mut DiagStyledString, len: usize, pos: usize, )

Adds a , to the type representation only if it is appropriate.

fn cmp_fn_sig( &self, sig1: &PolyFnSig<'tcx>, sig2: &PolyFnSig<'tcx>, ) -> (DiagStyledString, DiagStyledString)

Given two fn signatures highlight only sub-parts that are different.

pub fn cmp( &self, t1: Ty<'tcx>, t2: Ty<'tcx>, ) -> (DiagStyledString, DiagStyledString)

Compares two given types, eliding parts that are the same between them and highlighting relevant differences, and return two representation of those types for highlighted printing.

pub fn note_type_err( &self, diag: &mut Diag<'_>, cause: &ObligationCause<'tcx>, secondary_span: Option<(Span, Cow<'static, str>)>, values: Option<ValuePairs<'tcx>>, terr: TypeError<'tcx>, swap_secondary_and_primary: bool, prefer_label: bool, )

Extend a type error with extra labels pointing at “non-trivial” types, like closures and the return type of async fns.

secondary_span gives the caller the opportunity to expand diag with a span_label.

swap_secondary_and_primary is used to make projection errors in particular nicer by using the message in secondary_span as the primary label, and apply the message that would otherwise be used for the primary label on the secondary_span Span. This applies on E0271, like tests/ui/issues/issue-39970.stderr.

pub fn type_error_additional_suggestions( &self, trace: &TypeTrace<'tcx>, terr: TypeError<'tcx>, ) -> Vec<TypeErrorAdditionalDiags>

fn suggest_specify_actual_length( &self, terr: TypeError<'_>, trace: &TypeTrace<'_>, span: Span, ) -> Option<TypeErrorAdditionalDiags>

pub fn report_and_explain_type_error( &self, trace: TypeTrace<'tcx>, terr: TypeError<'tcx>, ) -> Diag<'a>

fn suggest_wrap_to_build_a_tuple( &self, span: Span, found: Ty<'tcx>, expected_fields: &List<Ty<'tcx>>, ) -> Option<TypeErrorAdditionalDiags>

fn values_str( &self, values: ValuePairs<'tcx>, ) -> Option<(DiagStyledString, DiagStyledString, Option<PathBuf>)>

fn expected_found_str_term( &self, exp_found: ExpectedFound<Term<'tcx>>, ) -> Option<(DiagStyledString, DiagStyledString, Option<PathBuf>)>

fn expected_found_str<T: Display + TypeFoldable<TyCtxt<'tcx>>>( &self, exp_found: ExpectedFound<T>, ) -> Option<(DiagStyledString, DiagStyledString, Option<PathBuf>)>

Returns a string of the form “expected {}, found {}”.

pub fn is_try_conversion(&self, span: Span, trait_def_id: DefId) -> bool

Determine whether an error associated with the given span and definition should be treated as being caused by the implicit From conversion within ? desugaring.

pub fn same_type_modulo_infer<T: Relate<TyCtxt<'tcx>>>( &self, a: T, b: T, ) -> bool

Structurally compares two types, modulo any inference variables.

Returns true if two types are equal, or if one type is an inference variable compatible with the other type. A TyVar inference type is compatible with any type, and an IntVar or FloatVar inference type are compatible with themselves or their concrete types (Int and Float types, respectively). When comparing two ADTs, these rules apply recursively.

impl<'a, 'tcx> TypeErrCtxt<'a, 'tcx>

pub(super) fn maybe_report_ambiguity( &self, obligation: &PredicateObligation<'tcx>, ) -> ErrorGuaranteed

fn annotate_source_of_ambiguity( &self, err: &mut Diag<'_>, ambiguities: &[CandidateSource], predicate: Predicate<'tcx>, )

impl<'a, 'tcx> TypeErrCtxt<'a, 'tcx>

pub fn report_selection_error( &self, obligation: PredicateObligation<'tcx>, root_obligation: &PredicateObligation<'tcx>, error: &SelectionError<'tcx>, ) -> ErrorGuaranteed

The root_obligation parameter should be the root_obligation field from a FulfillmentError. If no FulfillmentError is available, then it should be the same as obligation.

impl<'a, 'tcx> TypeErrCtxt<'a, 'tcx>

pub(super) fn apply_do_not_recommend( &self, obligation: &mut PredicateObligation<'tcx>, ) -> bool

fn emit_specialized_closure_kind_error( &self, obligation: &PredicateObligation<'tcx>, trait_ref: PolyTraitRef<'tcx>, ) -> Option<ErrorGuaranteed>

fn fn_arg_obligation( &self, obligation: &PredicateObligation<'tcx>, ) -> Result<(), ErrorGuaranteed>

fn try_conversion_context( &self, obligation: &PredicateObligation<'tcx>, trait_ref: TraitRef<'tcx>, err: &mut Diag<'_>, ) -> bool

When the E of the resulting Result<T, E> in an expression foo().bar().baz()?, identify those method chain sub-expressions that could or could not have been annotated with ?.

fn report_const_param_not_wf( &self, ty: Ty<'tcx>, obligation: &PredicateObligation<'tcx>, ) -> Diag<'a>

impl<'a, 'tcx> TypeErrCtxt<'a, 'tcx>

fn can_match_trait( &self, goal: TraitPredicate<'tcx>, assumption: PolyTraitPredicate<'tcx>, ) -> bool

fn can_match_projection( &self, goal: ProjectionPredicate<'tcx>, assumption: PolyProjectionPredicate<'tcx>, ) -> bool

pub(super) fn error_implies( &self, cond: Predicate<'tcx>, error: Predicate<'tcx>, ) -> bool

pub(super) fn report_projection_error( &self, obligation: &PredicateObligation<'tcx>, error: &MismatchedProjectionTypes<'tcx>, ) -> ErrorGuaranteed

fn maybe_detailed_projection_msg( &self, projection_term: AliasTerm<'tcx>, normalized_ty: Term<'tcx>, expected_ty: Term<'tcx>, ) -> Option<String>

pub fn fuzzy_match_tys( &self, a: Ty<'tcx>, b: Ty<'tcx>, ignoring_lifetimes: bool, ) -> Option<CandidateSimilarity>

pub(super) fn describe_closure(&self, kind: ClosureKind) -> &'static str

pub(super) fn find_similar_impl_candidates( &self, trait_pred: PolyTraitPredicate<'tcx>, ) -> Vec<ImplCandidate<'tcx>>

pub(super) fn report_similar_impl_candidates( &self, impl_candidates: &[ImplCandidate<'tcx>], trait_ref: PolyTraitRef<'tcx>, body_def_id: LocalDefId, err: &mut Diag<'_>, other: bool, param_env: ParamEnv<'tcx>, ) -> bool

fn report_similar_impl_candidates_for_root_obligation( &self, obligation: &PredicateObligation<'tcx>, trait_predicate: Binder<'tcx, TraitPredicate<'tcx>>, body_def_id: LocalDefId, err: &mut Diag<'_>, )

fn get_parent_trait_ref( &self, code: &ObligationCauseCode<'tcx>, ) -> Option<(Ty<'tcx>, Option<Span>)>

Gets the parent trait chain start

fn note_version_mismatch( &self, err: &mut Diag<'_>, trait_ref: PolyTraitRef<'tcx>, ) -> bool

If the Self type of the unsatisfied trait trait_ref implements a trait with the same path as trait_ref, a help message about a probable version mismatch is added to err

pub(super) fn mk_trait_obligation_with_new_self_ty( &self, param_env: ParamEnv<'tcx>, trait_ref_and_ty: Binder<'tcx, (TraitPredicate<'tcx>, Ty<'tcx>)>, ) -> PredicateObligation<'tcx>

Creates a PredicateObligation with new_self_ty replacing the existing type in the trait_ref.

For this to work, new_self_ty must have no escaping bound variables.

fn predicate_can_apply( &self, param_env: ParamEnv<'tcx>, pred: PolyTraitPredicate<'tcx>, ) -> bool

Returns true if the trait predicate may apply for some assignment to the type parameters.

pub fn note_obligation_cause( &self, err: &mut Diag<'_>, obligation: &PredicateObligation<'tcx>, )

pub(super) fn is_recursive_obligation( &self, obligated_types: &mut Vec<Ty<'tcx>>, cause_code: &ObligationCauseCode<'tcx>, ) -> bool

fn get_standard_error_message( &self, trait_predicate: PolyTraitPredicate<'tcx>, message: Option<String>, predicate_is_const: bool, append_const_msg: Option<AppendConstMessage>, post_message: String, ) -> String

fn get_safe_transmute_error_and_reason( &self, obligation: PredicateObligation<'tcx>, trait_ref: PolyTraitRef<'tcx>, span: Span, ) -> GetSafeTransmuteErrorAndReason

fn add_tuple_trait_message( &self, obligation_cause_code: &ObligationCauseCode<'tcx>, err: &mut Diag<'_>, )

fn try_to_add_help_message( &self, obligation: &PredicateObligation<'tcx>, trait_predicate: PolyTraitPredicate<'tcx>, err: &mut Diag<'_>, span: Span, is_fn_trait: bool, suggested: bool, unsatisfied_const: bool, )

fn add_help_message_for_fn_trait( &self, trait_ref: PolyTraitRef<'tcx>, err: &mut Diag<'_>, implemented_kind: ClosureKind, params: Binder<'tcx, Ty<'tcx>>, )

fn maybe_add_note_for_unsatisfied_const( &self, _trait_predicate: PolyTraitPredicate<'tcx>, _err: &mut Diag<'_>, _span: Span, ) -> UnsatisfiedConst

fn report_closure_error( &self, obligation: &PredicateObligation<'tcx>, closure_def_id: DefId, found_kind: ClosureKind, kind: ClosureKind, trait_prefix: &'static str, ) -> Diag<'a>

fn report_cyclic_signature_error( &self, obligation: &PredicateObligation<'tcx>, found_trait_ref: TraitRef<'tcx>, expected_trait_ref: TraitRef<'tcx>, terr: TypeError<'tcx>, ) -> Diag<'a>

fn report_opaque_type_auto_trait_leakage( &self, obligation: &PredicateObligation<'tcx>, def_id: DefId, ) -> ErrorGuaranteed

fn report_signature_mismatch_error( &self, obligation: &PredicateObligation<'tcx>, span: Span, found_trait_ref: TraitRef<'tcx>, expected_trait_ref: TraitRef<'tcx>, ) -> Result<Diag<'a>, ErrorGuaranteed>

pub fn get_fn_like_arguments( &self, node: Node<'_>, ) -> Option<(Span, Option<Span>, Vec<ArgKind>)>

Given some node representing a fn-like thing in the HIR map, returns a span and ArgKind information that describes the arguments it expects. This can be supplied to report_arg_count_mismatch.

pub fn report_arg_count_mismatch( &self, span: Span, found_span: Option<Span>, expected_args: Vec<ArgKind>, found_args: Vec<ArgKind>, is_closure: bool, closure_arg_span: Option<Span>, ) -> Diag<'a>

Reports an error when the number of arguments needed by a trait match doesn’t match the number that the expression provides.

pub fn type_implements_fn_trait( &self, param_env: ParamEnv<'tcx>, ty: Binder<'tcx, Ty<'tcx>>, polarity: PredicatePolarity, ) -> Result<(ClosureKind, Binder<'tcx, Ty<'tcx>>), ()>

Checks if the type implements one of Fn, FnMut, or FnOnce in that order, and returns the generic type corresponding to the argument of that trait (corresponding to the closure arguments).

fn report_not_const_evaluatable_error( &self, obligation: &PredicateObligation<'tcx>, span: Span, ) -> Result<Diag<'a>, ErrorGuaranteed>

impl<'tcx> TypeErrCtxt<'_, 'tcx>

fn impl_similar_to( &self, trait_ref: PolyTraitRef<'tcx>, obligation: &PredicateObligation<'tcx>, ) -> Option<(DefId, GenericArgsRef<'tcx>)>

fn describe_enclosure(&self, def_id: LocalDefId) -> Option<&'static str>

Used to set on_unimplemented’s ItemContext to be the enclosing (async) block/function/closure

pub fn on_unimplemented_note( &self, trait_ref: PolyTraitRef<'tcx>, obligation: &PredicateObligation<'tcx>, long_ty_file: &mut Option<PathBuf>, ) -> OnUnimplementedNote

impl<'a, 'tcx> TypeErrCtxt<'a, 'tcx>

pub fn report_overflow_error( &self, cause: OverflowCause<'tcx>, span: Span, suggest_increasing_limit: bool, mutate: impl FnOnce(&mut Diag<'_>), ) -> !

Reports that an overflow has occurred and halts compilation. We halt compilation unconditionally because it is important that overflows never be masked – they basically represent computations whose result could not be truly determined and thus we can’t say if the program type checks or not – and they are unusual occurrences in any case.

pub fn build_overflow_error( &self, cause: OverflowCause<'tcx>, span: Span, suggest_increasing_limit: bool, ) -> Diag<'a>

pub fn report_overflow_obligation<T>( &self, obligation: &Obligation<'tcx, T>, suggest_increasing_limit: bool, ) -> !
where T: Upcast<TyCtxt<'tcx>, Predicate<'tcx>> + Clone,

Reports that an overflow has occurred and halts compilation. We halt compilation unconditionally because it is important that overflows never be masked – they basically represent computations whose result could not be truly determined and thus we can’t say if the program type checks or not – and they are unusual occurrences in any case.

pub fn report_overflow_obligation_cycle( &self, cycle: &[PredicateObligation<'tcx>], ) -> !

Reports that a cycle was detected which led to overflow and halts compilation. This is equivalent to report_overflow_obligation except that we can give a more helpful error message (and, in particular, we do not suggest increasing the overflow limit, which is not going to help).

pub fn report_overflow_no_abort( &self, obligation: PredicateObligation<'tcx>, suggest_increasing_limit: bool, ) -> ErrorGuaranteed

impl<'a, 'tcx> TypeErrCtxt<'a, 'tcx>

pub fn suggest_restricting_param_bound( &self, err: &mut Diag<'_>, trait_pred: PolyTraitPredicate<'tcx>, associated_ty: Option<(&'static str, Ty<'tcx>)>, body_id: LocalDefId, )

pub(super) fn suggest_dereferences( &self, obligation: &PredicateObligation<'tcx>, err: &mut Diag<'_>, trait_pred: PolyTraitPredicate<'tcx>, ) -> bool

When after several dereferencing, the reference satisfies the trait bound. This function provides dereference suggestion for this specific situation.

fn get_closure_name( &self, def_id: DefId, err: &mut Diag<'_>, msg: Cow<'static, str>, ) -> Option<Symbol>

Given a closure’s DefId, return the given name of the closure.

This doesn’t account for reassignments, but it’s only used for suggestions.

pub(super) fn suggest_fn_call( &self, obligation: &PredicateObligation<'tcx>, err: &mut Diag<'_>, trait_pred: PolyTraitPredicate<'tcx>, ) -> bool

We tried to apply the bound to an fn or closure. Check whether calling it would evaluate to a type that would satisfy the trait bound. If it would, suggest calling it: bar(foo) → bar(foo()). This case is very likely to be hit if foo is async.

pub(super) fn check_for_binding_assigned_block_without_tail_expression( &self, obligation: &PredicateObligation<'tcx>, err: &mut Diag<'_>, trait_pred: PolyTraitPredicate<'tcx>, )

pub(super) fn suggest_add_clone_to_arg( &self, obligation: &PredicateObligation<'tcx>, err: &mut Diag<'_>, trait_pred: PolyTraitPredicate<'tcx>, ) -> bool

pub fn extract_callable_info( &self, body_id: LocalDefId, param_env: ParamEnv<'tcx>, found: Ty<'tcx>, ) -> Option<(DefIdOrName, Ty<'tcx>, Vec<Ty<'tcx>>)>

Extracts information about a callable type for diagnostics. This is a heuristic – it doesn’t necessarily mean that a type is always callable, because the callable type must also be well-formed to be called.

pub(super) fn suggest_add_reference_to_arg( &self, obligation: &PredicateObligation<'tcx>, err: &mut Diag<'_>, poly_trait_pred: PolyTraitPredicate<'tcx>, has_custom_message: bool, ) -> bool

pub(super) fn suggest_borrowing_for_object_cast( &self, err: &mut Diag<'_>, obligation: &PredicateObligation<'tcx>, self_ty: Ty<'tcx>, target_ty: Ty<'tcx>, )

pub(super) fn suggest_remove_reference( &self, obligation: &PredicateObligation<'tcx>, err: &mut Diag<'_>, trait_pred: PolyTraitPredicate<'tcx>, ) -> bool

Whenever references are used by mistake, like for (i, e) in &vec.iter().enumerate(), suggest removing these references until we reach a type that implements the trait.

pub(super) fn suggest_remove_await( &self, obligation: &PredicateObligation<'tcx>, err: &mut Diag<'_>, )

pub(super) fn suggest_change_mut( &self, obligation: &PredicateObligation<'tcx>, err: &mut Diag<'_>, trait_pred: PolyTraitPredicate<'tcx>, )

Check if the trait bound is implemented for a different mutability and note it in the final error.

pub(super) fn suggest_semicolon_removal( &self, obligation: &PredicateObligation<'tcx>, err: &mut Diag<'_>, span: Span, trait_pred: PolyTraitPredicate<'tcx>, ) -> bool

pub(super) fn return_type_span( &self, obligation: &PredicateObligation<'tcx>, ) -> Option<Span>

pub(super) fn suggest_impl_trait( &self, err: &mut Diag<'_>, obligation: &PredicateObligation<'tcx>, trait_pred: PolyTraitPredicate<'tcx>, ) -> bool

If all conditions are met to identify a returned dyn Trait, suggest using impl Trait if applicable and signal that the error has been expanded appropriately and needs to be emitted.

pub(super) fn point_at_returns_when_relevant( &self, err: &mut Diag<'_>, obligation: &PredicateObligation<'tcx>, )

pub(super) fn report_closure_arg_mismatch( &self, span: Span, found_span: Option<Span>, found: TraitRef<'tcx>, expected: TraitRef<'tcx>, cause: &ObligationCauseCode<'tcx>, found_node: Option<Node<'_>>, param_env: ParamEnv<'tcx>, ) -> Diag<'a>

fn note_conflicting_fn_args( &self, err: &mut Diag<'_>, cause: &ObligationCauseCode<'tcx>, expected: Ty<'tcx>, found: Ty<'tcx>, param_env: ParamEnv<'tcx>, )

fn note_conflicting_closure_bounds( &self, cause: &ObligationCauseCode<'tcx>, err: &mut Diag<'_>, )

pub(super) fn suggest_fully_qualified_path( &self, err: &mut Diag<'_>, item_def_id: DefId, span: Span, trait_ref: DefId, )

pub fn maybe_note_obligation_cause_for_async_await<G: EmissionGuarantee>( &self, err: &mut Diag<'_, G>, obligation: &PredicateObligation<'tcx>, ) -> bool

Adds an async-await specific note to the diagnostic when the future does not implement an auto trait because of a captured type.

note: future does not implement `Qux` as this value is used across an await
  --> $DIR/issue-64130-3-other.rs:17:5
   |
LL |     let x = Foo;
   |         - has type `Foo`
LL |     baz().await;
   |     ^^^^^^^^^^^ await occurs here, with `x` maybe used later
LL | }
   | - `x` is later dropped here

When the diagnostic does not implement Send or Sync specifically, then the diagnostic is “replaced” with a different message and a more specific error.

error: future cannot be sent between threads safely
  --> $DIR/issue-64130-2-send.rs:21:5
   |
LL | fn is_send<T: Send>(t: T) { }
   |               ---- required by this bound in `is_send`
...
LL |     is_send(bar());
   |     ^^^^^^^ future returned by `bar` is not send
   |
   = help: within `impl std::future::Future`, the trait `std::marker::Send` is not
           implemented for `Foo`
note: future is not send as this value is used across an await
  --> $DIR/issue-64130-2-send.rs:15:5
   |
LL |     let x = Foo;
   |         - has type `Foo`
LL |     baz().await;
   |     ^^^^^^^^^^^ await occurs here, with `x` maybe used later
LL | }
   | - `x` is later dropped here

Returns true if an async-await specific note was added to the diagnostic.

fn note_obligation_cause_for_async_await<G: EmissionGuarantee>( &self, err: &mut Diag<'_, G>, interior_or_upvar_span: CoroutineInteriorOrUpvar, is_async: bool, outer_coroutine: Option<DefId>, trait_pred: TraitPredicate<'tcx>, target_ty: Ty<'tcx>, obligation: &PredicateObligation<'tcx>, next_code: Option<&ObligationCauseCode<'tcx>>, )

Unconditionally adds the diagnostic note described in maybe_note_obligation_cause_for_async_await’s documentation comment.

pub(super) fn note_obligation_cause_code<G: EmissionGuarantee, T>( &self, body_id: LocalDefId, err: &mut Diag<'_, G>, predicate: T, param_env: ParamEnv<'tcx>, cause_code: &ObligationCauseCode<'tcx>, obligated_types: &mut Vec<Ty<'tcx>>, seen_requirements: &mut FxHashSet<DefId>, )
where T: Upcast<TyCtxt<'tcx>, Predicate<'tcx>>,

pub(super) fn suggest_await_before_try( &self, err: &mut Diag<'_>, obligation: &PredicateObligation<'tcx>, trait_pred: PolyTraitPredicate<'tcx>, span: Span, )

pub(super) fn suggest_floating_point_literal( &self, obligation: &PredicateObligation<'tcx>, err: &mut Diag<'_>, trait_ref: PolyTraitRef<'tcx>, )

pub fn suggest_derive( &self, obligation: &PredicateObligation<'tcx>, err: &mut Diag<'_>, trait_pred: PolyTraitPredicate<'tcx>, )

pub(super) fn suggest_dereferencing_index( &self, obligation: &PredicateObligation<'tcx>, err: &mut Diag<'_>, trait_pred: PolyTraitPredicate<'tcx>, )

fn note_function_argument_obligation<G: EmissionGuarantee>( &self, body_id: LocalDefId, err: &mut Diag<'_, G>, arg_hir_id: HirId, parent_code: &ObligationCauseCode<'tcx>, param_env: ParamEnv<'tcx>, failed_pred: Predicate<'tcx>, call_hir_id: HirId, )

fn suggest_option_method_if_applicable<G: EmissionGuarantee>( &self, failed_pred: Predicate<'tcx>, param_env: ParamEnv<'tcx>, err: &mut Diag<'_, G>, expr: &Expr<'_>, )

fn look_for_iterator_item_mistakes<G: EmissionGuarantee>( &self, assocs_in_this_method: &[Option<(Span, (DefId, Ty<'tcx>))>], typeck_results: &TypeckResults<'tcx>, type_diffs: &[TypeError<'tcx>], param_env: ParamEnv<'tcx>, path_segment: &PathSegment<'_>, args: &[Expr<'_>], err: &mut Diag<'_, G>, )

fn point_at_chain<G: EmissionGuarantee>( &self, expr: &Expr<'_>, typeck_results: &TypeckResults<'tcx>, type_diffs: Vec<TypeError<'tcx>>, param_env: ParamEnv<'tcx>, err: &mut Diag<'_, G>, )

fn probe_assoc_types_at_expr( &self, type_diffs: &[TypeError<'tcx>], span: Span, prev_ty: Ty<'tcx>, body_id: HirId, param_env: ParamEnv<'tcx>, ) -> Vec<Option<(Span, (DefId, Ty<'tcx>))>>

pub(super) fn suggest_convert_to_slice( &self, err: &mut Diag<'_>, obligation: &PredicateObligation<'tcx>, trait_ref: PolyTraitRef<'tcx>, candidate_impls: &[ImplCandidate<'tcx>], span: Span, )

If the type that failed selection is an array or a reference to an array, but the trait is implemented for slices, suggest that the user converts the array into a slice.

pub(super) fn explain_hrtb_projection( &self, diag: &mut Diag<'_>, pred: PolyTraitPredicate<'tcx>, param_env: ParamEnv<'tcx>, cause: &ObligationCause<'tcx>, )

pub(super) fn suggest_desugaring_async_fn_in_trait( &self, err: &mut Diag<'_>, trait_ref: PolyTraitRef<'tcx>, )

pub fn ty_kind_suggestion( &self, param_env: ParamEnv<'tcx>, ty: Ty<'tcx>, ) -> Option<String>

pub(super) fn suggest_add_result_as_return_type( &self, obligation: &PredicateObligation<'tcx>, err: &mut Diag<'_>, trait_pred: PolyTraitPredicate<'tcx>, )

pub(super) fn suggest_unsized_bound_if_applicable( &self, err: &mut Diag<'_>, obligation: &PredicateObligation<'tcx>, )

fn suggest_indirection_for_unsized( &self, err: &mut Diag<'_>, item: &Item<'tcx>, param: &GenericParam<'tcx>, ) -> bool

impl<'a, 'tcx> TypeErrCtxt<'a, 'tcx>

pub fn report_fulfillment_errors( &self, errors: Vec<FulfillmentError<'tcx>>, ) -> ErrorGuaranteed

fn report_fulfillment_error( &self, error: &FulfillmentError<'tcx>, ) -> ErrorGuaranteed

impl<'a, 'tcx> TypeErrCtxt<'a, 'tcx>

pub fn report_extra_impl_obligation( &self, error_span: Span, impl_item_def_id: LocalDefId, trait_item_def_id: DefId, requirement: &dyn Display, ) -> Diag<'a>

impl<'a, 'tcx> TypeErrCtxt<'a, 'tcx>

pub fn dcx(&self) -> DiagCtxtHandle<'a>

pub fn err_ctxt(&self) -> !

👎Deprecated: you already have a TypeErrCtxt

This is just to avoid a potential footgun of accidentally dropping typeck_results by calling InferCtxt::err_ctxt