rustc_trait_selection::error_reporting::traits

Function attempt_dyn_to_enum_suggestion

fn attempt_dyn_to_enum_suggestion(
    tcx: TyCtxt<'_>,
    trait_def_id: DefId,
    trait_str: &str,
    err: &mut Diag<'_>,
)

Attempt to suggest converting the dyn Trait argument to an enumeration over the types that implement Trait.