rustc_trait_selection::error_reporting::traits

Function attempt_dyn_to_impl_suggestion

fn attempt_dyn_to_impl_suggestion(
    tcx: TyCtxt<'_>,
    hir_id: Option<HirId>,
    err: &mut Diag<'_>,
)

Attempt to suggest that a dyn Trait argument or return type be converted to use impl Trait.