Module rustc_trait_selection::infer::canonical::ir::solve

Modules

inspect

Data structure used to inspect trait solver behavior.

Structs

CacheData

ExternalConstraintsData

Additional constraints returned on success.

Goal

A goal is a statement, i.e. predicate, we want to prove given some assumptions, i.e. param_env.

NestedNormalizationGoals

NoSolution

PredefinedOpaquesData

Opaques that are defined in the inference context before a query is called.

QueryInput

Response

Enums

BuiltinImplSource

CandidateSource

Possible ways the given goal can be proven.

Certainty

GoalSource

Why a specific goal has to be proven.

MaybeCause

Why we failed to evaluate a goal.

Reveal

Depending on the stage of compilation, we want projection to be more or less conservative.

SolverMode

Type Aliases

CanonicalInput

CanonicalResponse

QueryResult

The result of evaluating a canonical query.