rustc_trait_selection::infer
Derive Macro TypeVisitable

#[derive(TypeVisitable)]
{
    // Attributes available to this derive:
    #[type_visitable]
}

Derives TypeVisitable for the annotated struct or enum (union is not supported).

Each field of the struct or enum variant will be visited in definition order, using the TypeVisitable implementation for its type. However, if a field of a struct or an enum variant is annotated with #[type_visitable(ignore)] then that field will not be visited (and its type is not required to implement TypeVisitable).