Enum rustc_trait_selection::infer::RegionVariableOrigin

pub enum RegionVariableOrigin {
    MiscVariable(Span),
    PatternRegion(Span),
    AddrOfRegion(Span),
    Autoref(Span),
    Coercion(Span),
    EarlyBoundRegion(Span, Symbol),
    LateBoundRegion(Span, BoundRegionKind, LateBoundRegionConversionTime),
    UpvarRegion(UpvarId, Span),
    Nll(NllRegionVariableOrigin),
}

Reasons to create a region inference variable

See error_reporting module for more details

Variants

MiscVariable(Span)

Region variables created for ill-categorized reasons, mostly indicates places in need of refactoring

PatternRegion(Span)

Regions created by a &P or [...] pattern

AddrOfRegion(Span)

Regions created by & operator

Autoref(Span)

Regions created as part of an autoref of a method receiver

Coercion(Span)

Regions created as part of an automatic coercion

EarlyBoundRegion(Span, Symbol)

Region variables created as the values for early-bound regions

LateBoundRegion(Span, BoundRegionKind, LateBoundRegionConversionTime)

Region variables created for bound regions in a function or method that is called

UpvarRegion(UpvarId, Span)

Nll(NllRegionVariableOrigin)

This origin is used for the inference variables that we create during NLL region processing.