Module rustc_trait_selection::infer::error_reporting::nice_region_error

Modules

find_anon_type

Structs

HirTraitObjectVisitor

Collect all hir::Ty<'_> Spans for trait objects with an implicit lifetime.

NiceRegionError

TraitObjectVisitor

Collect all the trait objects in a type that could have received an implicit 'static lifetime.

Functions

find_anon_type

This function calls the visit_ty method for the parameters corresponding to the anonymous regions. The nested_visitor.found_type contains the anonymous type.

find_param_with_region

suggest_adding_lifetime_params

Currently only used in rustc_borrowck, probably should be removed in favour of public_errors::AddLifetimeParamsSuggestion

suggest_new_region_bound