Struct rustc_trait_selection::infer::error_reporting::TypeErrCtxt

pub struct TypeErrCtxt<'a, 'tcx> {
    pub infcx: &'a InferCtxt<'tcx>,
    pub typeck_results: Option<Ref<'a, TypeckResults<'tcx>>>,
    pub fallback_has_occurred: bool,
    pub normalize_fn_sig: Box<dyn Fn(Binder<'tcx, FnSig<'tcx>>) -> Binder<'tcx, FnSig<'tcx>> + 'a, Global>,
    pub autoderef_steps: Box<dyn Fn(Ty<'tcx>) -> Vec<(Ty<'tcx>, Vec<Obligation<'tcx, Predicate<'tcx>>, Global>), Global> + 'a, Global>,
}

A helper for building type related errors. The typeck_results field is only populated during an in-progress typeck. Get an instance by calling InferCtxt::err_ctxt or FnCtxt::err_ctxt.

You must only create this if you intend to actually emit an error. This provides a lot of utility methods which should not be used during the happy path.

Fields
infcx: &'a InferCtxt<'tcx>typeck_results: Option<Ref<'a, TypeckResults<'tcx>>>fallback_has_occurred: boolnormalize_fn_sig: Box<dyn Fn(Binder<'tcx, FnSig<'tcx>>) -> Binder<'tcx, FnSig<'tcx>> + 'a, Global>autoderef_steps: Box<dyn Fn(Ty<'tcx>) -> Vec<(Ty<'tcx>, Vec<Obligation<'tcx, Predicate<'tcx>>, Global>), Global> + 'a, Global>