Type Definition rustc_trait_selection::infer::outlives::env::RegionBoundPairs

pub type RegionBoundPairs<'tcx> = IndexSet<OutlivesPredicate<GenericKind<'tcx>, Region<'tcx>>, BuildHasherDefault<FxHasher>>;

“Region-bound pairs” tracks outlives relations that are known to be true, either because of explicit where-clauses like T: 'a or because of implied bounds.

Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference’s “Type Layout” chapter for details on type layout guarantees.

Size: 56 bytes