Enum rustc_trait_selection::infer::region_constraints::VerifyBound

pub enum VerifyBound<'tcx> {
    IfEq(Binder<'tcx, VerifyIfEq<'tcx>>),
    OutlivedBy(Region<'tcx>),
    IsEmpty,
    AnyBound(Vec<VerifyBound<'tcx>, Global>),
    AllBounds(Vec<VerifyBound<'tcx>, Global>),
}

Describes the things that some GenericKind value G is known to outlive. Each variant of VerifyBound can be thought of as a function:

fn(min: Region) -> bool { .. }

where true means that the region min meets that G: min. (False means nothing.)

So, for example, if we have the type T and we have in scope that T: 'a and T: 'b, then the verify bound might be:

fn(min: Region) -> bool {
   ('a: min) || ('b: min)
}

This is described with an AnyRegion('a, 'b) node.

Variants

IfEq(Binder<'tcx, VerifyIfEq<'tcx>>)

See VerifyIfEq docs

OutlivedBy(Region<'tcx>)

Given a region R, expands to the function:

fn(min) -> bool {
    R: min
}

This is used when we can establish that G: R – therefore, if R: min, then by transitivity G: min.

IsEmpty

Given a region R, true if it is 'empty.

AnyBound(Vec<VerifyBound<'tcx>, Global>)

Given a set of bounds B, expands to the function:

fn(min) -> bool {
    exists (b in B) { b(min) }
}

In other words, if we meet some bound in B, that suffices. This is used when all the bounds in B are known to apply to G.

AllBounds(Vec<VerifyBound<'tcx>, Global>)

Given a set of bounds B, expands to the function:

fn(min) -> bool {
    forall (b in B) { b(min) }
}

In other words, if we meet all bounds in B, that suffices. This is used when some bound in B is known to suffice, but we don’t know which.