rustc_trait_selection::infer
Trait InferCtxtBuilderExt

pub trait InferCtxtBuilderExt<'tcx> {
    // Required method
    fn enter_canonical_trait_query<K, R>(
        self,
        canonical_key: &CanonicalQueryInput<'tcx, K>,
        operation: impl FnOnce(&ObligationCtxt<'_, 'tcx>, K) -> Result<R, NoSolution>,
    ) -> Result<CanonicalQueryResponse<'tcx, R>, NoSolution>
       where K: TypeFoldable<TyCtxt<'tcx>>,
             R: Debug + TypeFoldable<TyCtxt<'tcx>>,
             Canonical<'tcx, QueryResponse<'tcx, R>>: ArenaAllocatable<'tcx>;
}

Required Methods

fn enter_canonical_trait_query<K, R>( self, canonical_key: &CanonicalQueryInput<'tcx, K>, operation: impl FnOnce(&ObligationCtxt<'_, 'tcx>, K) -> Result<R, NoSolution>, ) -> Result<CanonicalQueryResponse<'tcx, R>, NoSolution>
where K: TypeFoldable<TyCtxt<'tcx>>, R: Debug + TypeFoldable<TyCtxt<'tcx>>, Canonical<'tcx, QueryResponse<'tcx, R>>: ArenaAllocatable<'tcx>,

The “main method” for a canonicalized trait query. Given the canonical key canonical_key, this method will create a new inference context, instantiate the key, and run your operation op. The operation should yield up a result (of type R) as well as a set of trait obligations that must be fully satisfied. These obligations will be processed and the canonical result created.

Returns NoSolution in the event of any error.

(It might be mildly nicer to implement this on TyCtxt, and not InferCtxtBuilder, but that is a bit tricky right now. In part because we would need a for<'tcx> sort of bound for the closure and in part because it is convenient to have 'tcx be free on this function so that we can talk about K: TypeFoldable<TyCtxt<'tcx>>.)

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.