rustc_trait_selection::regions
Trait InferCtxtRegionExt

pub trait InferCtxtRegionExt<'tcx> {
    // Required method
    fn resolve_regions(
        &self,
        outlives_env: &OutlivesEnvironment<'tcx>,
    ) -> Vec<RegionResolutionError<'tcx>>;
}

Required Methods

fn resolve_regions( &self, outlives_env: &OutlivesEnvironment<'tcx>, ) -> Vec<RegionResolutionError<'tcx>>

Resolve regions, using the deep normalizer to normalize any type-outlives obligations in the process. This is in rustc_trait_selection because we need to normalize.

Prefer this method over resolve_regions_with_normalize, unless you are doing something specific for normalization.