rustc_trait_selection

Module solve

Modules

delegate 🔒

fulfill 🔒

inspect

normalize 🔒

select 🔒

Structs

EvalCtxt

ExternalConstraintsData

Additional constraints returned on success.

FulfillmentCtxt

A trait engine using the new trait solver.

Goal

A goal is a statement, i.e. predicate, we want to prove given some assumptions, i.e. param_env.

GoalEvaluation

The result of evaluating a goal.

GoalStalledOn

The conditions that must change for a goal to warrant

NestedNormalizationGoals

NoSolution

PredefinedOpaquesData

Opaques that are defined in the inference context before a query is called.

QueryInput

Response

StalledOnCoroutines

Detect if a goal is stalled on a coroutine that is owned by the current typeck root.

Enums

AdtDestructorKind

Indicates that a impl Drop for Adt is const or not.

BuiltinImplSource

CandidateSource

Possible ways the given goal can be proven.

Certainty

GenerateProofTree

GoalSource

Why a specific goal has to be proven.

HasChanged

Whether evaluating this goal ended up changing the inference state.

MaybeCause

Why we failed to evaluate a goal.

NextSolverError

ParamEnvSource

SizedTraitKind

Which sizedness trait - Sized, MetaSized? PointeeSized is omitted as it is removed during lowering.

Traits

InferCtxtSelectExt

SolverDelegateEvalExt

Functions

deeply_normalize

Deeply normalize all aliases in value. This does not handle inference and expects its input to be already fully resolved.

deeply_normalize_with_skipped_universes

Deeply normalize all aliases in value. This does not handle inference and expects its input to be already fully resolved.

deeply_normalize_with_skipped_universes_and_ambiguous_coroutine_goals

Deeply normalize all aliases in value. This does not handle inference and expects its input to be already fully resolved.

Type Aliases

CanonicalInput

CanonicalResponse

QueryResult

The result of evaluating a canonical query.