Module rustc_trait_selection::solve::inspect

Modules

analyse 🔒

build 🔒

Building proof trees incrementally during trait solving.

Structs

AddedGoalsEvaluation

CanonicalGoalEvaluation

GoalEvaluation

GoalEvaluationStep

InspectCandidate

InspectGoal

Probe

A self-contained computation during trait solving. This either corresponds to a EvalCtxt::probe(_X) call or the root evaluation of a goal.

State

Some data together with information about how they relate to the input of the canonical query.

Enums

CacheHit

CanonicalGoalEvaluationKind

GoalEvaluationKind

When evaluating the root goals we also store the original values for the CanonicalVarValues of the canonicalized goal. We use this to map any CanonicalState from the local InferCtxt of the solver query to the InferCtxt of the caller.

ProbeKind

What kind of probe we’re in. In case the probe represents a candidate, or the final result of the current goal - via ProbeKind::Root - we also store the QueryResult.

ProbeStep

Traits

ProofTreeInferCtxtExt

ProofTreeVisitor

The public API to interact with proof trees.

Type Aliases

CanonicalState