Struct rustc_trait_selection::solve::FulfillmentCtxt

pub struct FulfillmentCtxt<'tcx> {
    obligations: Vec<PredicateObligation<'tcx>>,
    usable_in_snapshot: usize,
}

A trait engine using the new trait solver.

This is mostly identical to how evaluate_all works inside of the solver, except that the requirements are slightly different.

Unlike evaluate_all it is possible to add new obligations later on and we also have to track diagnostics information by using Obligation instead of Goal.

It is also likely that we want to use slightly different datastructures here as this will have to deal with far more root goals than evaluate_all.

Fields
obligations: Vec<PredicateObligation<'tcx>>usable_in_snapshot: usize

The snapshot in which this context was created. Using the context outside of this snapshot leads to subtle bugs if the snapshot gets rolled back. Because of this we explicitly check that we only use the context in exactly this snapshot.

Implementations

impl<'tcx> FulfillmentCtxt<'tcx>

pub fn new(infcx: &InferCtxt<'tcx>) -> FulfillmentCtxt<'tcx>