Struct rustc_trait_selection::traits::engine::ObligationCtxt

pub struct ObligationCtxt<'a, 'tcx> {
    pub infcx: &'a InferCtxt<'tcx>,
    engine: RefCell<Box<dyn TraitEngine<'tcx>>>,
}

Used if you want to have pleasant experience when dealing with obligations outside of hir or mir typeck.

Fields
infcx: &'a InferCtxt<'tcx>engine: RefCell<Box<dyn TraitEngine<'tcx>>>
Implementations

impl<'a, 'tcx> ObligationCtxt<'a, 'tcx>

pub fn new(infcx: &'a InferCtxt<'tcx>) -> Self

pub fn new_in_snapshot(infcx: &'a InferCtxt<'tcx>) -> Self

pub fn register_obligation(&self, obligation: PredicateObligation<'tcx>)

pub fn register_obligations(
&self,
obligations: impl IntoIterator<Item = PredicateObligation<'tcx>>
)

pub fn register_infer_ok_obligations<T>(&self, infer_ok: InferOk<'tcx, T>) -> T

pub fn register_bound(
&self,
cause: ObligationCause<'tcx>,
param_env: ParamEnv<'tcx>,
ty: Ty<'tcx>,
def_id: DefId
)

Requires that ty must implement the trait with def_id in the given environment. This trait must not have any type parameters (except for Self).

pub fn normalize<T: TypeFoldable<'tcx>>(
&self,
cause: &ObligationCause<'tcx>,
param_env: ParamEnv<'tcx>,
value: T
) -> T

pub fn eq_exp<T>(
&self,
cause: &ObligationCause<'tcx>,
param_env: ParamEnv<'tcx>,
a_is_expected: bool,
a: T,
b: T
) -> Result<(), TypeError<'tcx>>where
T: ToTrace<'tcx>,

Makes expected <: actual.

pub fn eq<T: ToTrace<'tcx>>(
&self,
cause: &ObligationCause<'tcx>,
param_env: ParamEnv<'tcx>,
expected: T,
actual: T
) -> Result<(), TypeError<'tcx>>

pub fn sub<T: ToTrace<'tcx>>(
&self,
cause: &ObligationCause<'tcx>,
param_env: ParamEnv<'tcx>,
expected: T,
actual: T
) -> Result<(), TypeError<'tcx>>

Checks whether expected is a subtype of actual: expected <: actual.

pub fn sup<T: ToTrace<'tcx>>(
&self,
cause: &ObligationCause<'tcx>,
param_env: ParamEnv<'tcx>,
expected: T,
actual: T
) -> Result<(), TypeError<'tcx>>

Checks whether expected is a supertype of actual: expected :> actual.

pub fn select_where_possible(&self) -> Vec<FulfillmentError<'tcx>>

pub fn select_all_or_error(&self) -> Vec<FulfillmentError<'tcx>>

pub fn assumed_wf_types(
&self,
param_env: ParamEnv<'tcx>,
span: Span,
def_id: LocalDefId
) -> FxIndexSet<Ty<'tcx>>

pub fn make_canonicalized_query_response<T>(
&self,
inference_vars: CanonicalVarValues<'tcx>,
answer: T
) -> Fallible<CanonicalizedQueryResponse<'tcx, T>>where
T: Debug + TypeFoldable<'tcx>,
Canonical<'tcx, QueryResponse<'tcx, T>>: ArenaAllocatable<'tcx>,