Enum rustc_trait_selection::traits::SelectionError

pub enum SelectionError<'tcx> {
    Unimplemented,
    OutputTypeParameterMismatch(Box<SelectionOutputTypeParameterMismatch<'tcx>, Global>),
    TraitNotObjectSafe(DefId),
    NotConstEvaluatable(NotConstEvaluatable),
    Overflow(OverflowError),
    ErrorReporting,
    OpaqueTypeAutoTraitLeakageUnknown(DefId),
}

Variants

Unimplemented

The trait is not implemented.

OutputTypeParameterMismatch(Box<SelectionOutputTypeParameterMismatch<'tcx>, Global>)

After a closure impl has selected, its “outputs” were evaluated (which for closures includes the “input” type params) and they didn’t resolve. See confirm_poly_trait_refs for more.

TraitNotObjectSafe(DefId)

The trait pointed by DefId is not object safe.

NotConstEvaluatable(NotConstEvaluatable)

A given constant couldn’t be evaluated.

Overflow(OverflowError)

Exceeded the recursion depth during type projection.

ErrorReporting

Signaling that an error has already been emitted, to avoid multiple errors being shown.

OpaqueTypeAutoTraitLeakageUnknown(DefId)

Computing an opaque type’s hidden type caused an error (e.g. a cycle error). We can thus not know whether the hidden type implements an auto trait, so we should not presume anything about it.