Module rustc_trait_selection::traits::error_reporting

Modules

on_unimplemented

suggestions

Structs

FindTypeParam 🔒

Look for type param in an ADT being used only through a reference to confirm that suggesting param: ?Sized would be a valid constraint.

HasNumericInferVisitor 🔒

ImplCandidate

Enums

ArgKind

Summarizes information

CandidateSimilarity

Traits

InferCtxtExt

InferCtxtPrivExt 🔒

Functions

get_option_generic_from_field_id 🔒

Extract the span for the generic type T of Option<T> in a field definition

recursive_type_with_infinite_size_error

report_object_safety_error