Module rustc_trait_selection::traits::error_reporting::suggestions

Structs

AwaitsVisitor 🔒

Collect all the awaited expressions within the input expression.

ReplaceImplTraitFolder 🔒

ReplaceImplTraitVisitor 🔒

Collect the spans that we see the generic param param_did

ReturnsVisitor

Collect all the returned expressions within the input expression. Used to point at the return spans when we want to suggest some change to them.

Enums

GeneratorData

GeneratorInteriorOrUpvar

Traits

NextTypeParamName

TypeErrCtxtExt

Functions

hint_missing_borrow 🔒

Add a hint to add a missing borrow or remove an unnecessary one.

predicate_constraint 🔒

suggest_restriction 🔒

Type parameter needs more bounds. The trivial case is T where T: Bound, but it can also be an impl Trait param that needs to be decomposed to a type param for cleaner code.

suggest_trait_object_return_type_alternatives 🔒