Struct rustc_trait_selection::traits::fulfill::FulfillmentContext

pub struct FulfillmentContext<'tcx> {
    predicates: ObligationForest<PendingPredicateObligation<'tcx>>,
    relationships: FxHashMap<TyVid, FoundRelationships>,
    usable_in_snapshot: bool,
}

The fulfillment context is used to drive trait resolution. It consists of a list of obligations that must be (eventually) satisfied. The job is to track which are satisfied, which yielded errors, and which are still pending. At any point, users can call select_where_possible, and the fulfillment context will try to do selection, retaining only those obligations that remain ambiguous. This may be helpful in pushing type inference along. Once all type inference constraints have been generated, the method select_all_or_error can be used to report any remaining ambiguous cases as errors.

Fields
predicates: ObligationForest<PendingPredicateObligation<'tcx>>relationships: FxHashMap<TyVid, FoundRelationships>usable_in_snapshot: bool
Implementations

impl<'a, 'tcx> FulfillmentContext<'tcx>

pub fn new() -> FulfillmentContext<'tcx>

Creates a new fulfillment context.

pub fn new_in_snapshot() -> FulfillmentContext<'tcx>

fn select(
&mut self,
selcx: &mut SelectionContext<'a, 'tcx>
) -> Vec<FulfillmentError<'tcx>>

Attempts to select obligations using selcx.