rustc_trait_selection::traits
Module misc

Miscellaneous type-system utilities that are too small to deserve their own modules.

Enums

ConstParamTyImplementationError

CopyImplementationError

InfringingFieldsReason

Functions

all_fields_implement_trait

Check that all fields of a given adt implement lang_item trait.

type_allowed_to_implement_const_param_ty

Checks that the fields of the type (an ADT) all implement (Unsized?)ConstParamTy.

type_allowed_to_implement_copy

Checks that the fields of the type (an ADT) all implement copy.