rustc_trait_selection::traits::query::dropck_outlives
Function dtorck_constraint_for_ty_inner

pub fn dtorck_constraint_for_ty_inner<'tcx>(
    tcx: TyCtxt<'tcx>,
    param_env: ParamEnv<'tcx>,
    span: Span,
    depth: usize,
    ty: Ty<'tcx>,
    constraints: &mut DropckConstraint<'tcx>,
) -> Result<(), NoSolution>

Returns a set of constraints that needs to be satisfied in order for ty to be valid for destruction.