rustc_trait_selection::traits::query
Enum OutlivesBound

pub enum OutlivesBound<'tcx> {
    RegionSubRegion(Region<'tcx>, Region<'tcx>),
    RegionSubParam(Region<'tcx>, ParamTy),
    RegionSubAlias(Region<'tcx>, AliasTy<TyCtxt<'tcx>>),
}

Outlives bounds are relationships between generic parameters, whether they both be regions ('a: 'b) or whether types are involved (T: 'a). These relationships can be extracted from the full set of predicates we understand or also from types (in which case they are called implied bounds). They are fed to the OutlivesEnv which in turn is supplied to the region checker and other parts of the inference system.

Variants

RegionSubRegion(Region<'tcx>, Region<'tcx>)

RegionSubParam(Region<'tcx>, ParamTy)

RegionSubAlias(Region<'tcx>, AliasTy<TyCtxt<'tcx>>)