rustc_trait_selection::traits::query::type_op
Module implied_outlives_bounds

Structs

ImpliedOutlivesBounds

Functions

compute_implied_outlives_bounds_compat_inner

compute_implied_outlives_bounds_inner

implied_bounds_from_components 🔒

When we have an implied bound that T: 'a, we can further break this down to determine what relationships would have to hold for T: 'a to hold. We get to assume that the caller has validated those relationships.