Struct rustc_trait_selection::traits::select::TraitObligationStack

struct TraitObligationStack<'prev, 'tcx> {
    obligation: &'prev TraitObligation<'tcx>,
    fresh_trait_pred: PolyTraitPredicate<'tcx>,
    reached_depth: Cell<usize>,
    previous: TraitObligationStackList<'prev, 'tcx>,
    depth: usize,
    dfn: usize,
}

Fields
obligation: &'prev TraitObligation<'tcx>fresh_trait_pred: PolyTraitPredicate<'tcx>

The trait predicate from obligation but “freshened” with the selection-context’s freshener. Used to check for recursion.

reached_depth: Cell<usize>

Starts out equal to depth – if, during evaluation, we encounter a cycle, then we will set this flag to the minimum depth of that cycle for all participants in the cycle. These participants will then forego caching their results. This is not the most efficient solution, but it addresses #60010. The problem we are trying to prevent:

If you have A: AutoTrait requires B: AutoTrait and C: NonAutoTrait

B: AutoTrait requires A: AutoTrait (coinductive cycle, ok)

C: NonAutoTrait requires A: AutoTrait (non-coinductive cycle, not ok)

you don’t want to cache that B: AutoTrait or A: AutoTrait is EvaluatedToOk; this is because they were only considered ok on the premise that if A: AutoTrait held, but we indeed encountered a problem (later on) with A: AutoTrait. So we currently set a flag on the stack node for B: AutoTrait(as well as the second instance ofA: AutoTrait`) to suppress caching.

This is a simple, targeted fix. A more-performant fix requires deeper changes, but would permit more caching: we could basically defer caching until we have fully evaluated the tree, and then cache the entire tree at once. In any case, the performance impact here shouldn’t be so horrible: every time this is hit, we do cache at least one trait, so we only evaluate each member of a cycle up to N times, where N is the length of the cycle. This means the performance impact is bounded and we shouldn’t have any terrible worst-cases.

previous: TraitObligationStackList<'prev, 'tcx>depth: usize

The number of parent frames plus one (thus, the topmost frame has depth 1).

dfn: usize

The depth-first number of this node in the search graph – a pre-order index. Basically, a freshly incremented counter.

Implementations

impl<'o, 'tcx> TraitObligationStack<'o, 'tcx>

fn list(&'o self) -> TraitObligationStackList<'o, 'tcx>

fn cache(&self) -> &'o ProvisionalEvaluationCache<'tcx>

fn iter(&'o self) -> TraitObligationStackList<'o, 'tcx>

fn update_reached_depth(&self, reached_depth: usize)

Indicates that attempting to evaluate this stack entry required accessing something from the stack at depth reached_depth.