rustc_trait_selection::traits

Module solve

Modules

inspect

Data structure used to inspect trait solver behavior.

Structs

ExternalConstraints

ExternalConstraintsData

Additional constraints returned on success.

NestedNormalizationGoals

NoSolution

PredefinedOpaques

PredefinedOpaquesData

Opaques that are defined in the inference context before a query is called.

Response

Enums

AdtDestructorKind

Indicates that a impl Drop for Adt is const or not.

BuiltinImplSource

Certainty

GoalSource

Why a specific goal has to be proven.

MaybeCause

Why we failed to evaluate a goal.

ParamEnvSource

SizedTraitKind

Which sizedness trait - Sized, MetaSized? PointeeSized is omitted as it is removed during lowering.

Type Aliases

CandidateSource

CanonicalInput

CanonicalResponse

Goal

QueryInput

QueryResult