Enum rustc_trait_selection::traits::solve::Certainty

pub enum Certainty {
    Yes,
    Maybe(MaybeCause),
}

Variants

Yes

Maybe(MaybeCause)

Implementations

impl Certainty

pub const AMBIGUOUS: Certainty = _

pub fn unify_with(self, other: Certainty) -> Certainty

Use this function to merge the certainty of multiple nested subgoals.

Given an impl like impl<T: Foo + Bar> Baz for T {}, we have 2 nested subgoals whenever we use the impl as a candidate: T: Foo and T: Bar. If evaluating T: Foo results in ambiguity and T: Bar results in success, we merge these two responses. This results in ambiguity.

If we unify ambiguity with overflow, we return overflow. This doesnโt matter inside of the solver as we do not distinguish ambiguity from overflow. It does however matter for diagnostics. If T: Foo resulted in overflow and T: Bar in ambiguity without changing the inference state, we still want to tell the user that T: Baz results in overflow.

pub const fn overflow(suggest_increasing_limit: bool) -> Certainty