Trait rustc_trait_selection::traits::specialization_graph::GraphExt

pub trait GraphExt {
    fn insert(
        &mut self,
        tcx: TyCtxt<'_>,
        impl_def_id: DefId,
        overlap_mode: OverlapMode
    ) -> Result<Option<FutureCompatOverlapError>, OverlapError>;
    fn record_impl_from_cstore(
        &mut self,
        tcx: TyCtxt<'_>,
        parent: DefId,
        child: DefId
    );
}

Required Methods

fn insert(
&mut self,
tcx: TyCtxt<'_>,
impl_def_id: DefId,
overlap_mode: OverlapMode
) -> Result<Option<FutureCompatOverlapError>, OverlapError>

Insert a local impl into the specialization graph. If an existing impl conflicts with it (has overlap, but neither specializes the other), information about the area of overlap is returned in the Err.

fn record_impl_from_cstore(&mut self, tcx: TyCtxt<'_>, parent: DefId, child: DefId)

Insert cached metadata mapping from a child impl back to its parent.