Function rustc_trait_selection::traits::specialize::fulfill_implication

fn fulfill_implication<'a, 'tcx>(
    infcx: &InferCtxt<'a, 'tcx>,
    param_env: ParamEnv<'tcx>,
    source_trait_ref: TraitRef<'tcx>,
    target_impl: DefId
) -> Result<SubstsRef<'tcx>, ()>

Attempt to fulfill all obligations of target_impl after unification with source_trait_ref. If successful, returns a substitution for all the generics of target_impl, including both those needed to unify with source_trait_ref and those whose identity is determined via a where clause in the impl.