rustc_trait_selection::traits
Struct Obligation

pub struct Obligation<'tcx, T> {
    pub cause: ObligationCause<'tcx>,
    pub param_env: ParamEnv<'tcx>,
    pub predicate: T,
    pub recursion_depth: usize,
}

An Obligation represents some trait reference (e.g., i32: Eq) for which the “impl_source” must be found. The process of finding an “impl_source” is called “resolving” the Obligation. This process consists of either identifying an impl (e.g., impl Eq for i32) that satisfies the obligation, or else finding a bound that is in scope. The eventual result is usually a Selection (defined below).

Fields
cause: ObligationCause<'tcx>

The reason we have to prove this thing.

param_env: ParamEnv<'tcx>

The environment in which we should prove this thing.

predicate: T

The thing we are trying to prove.

recursion_depth: usize

If we started proving this as a result of trying to prove something else, track the total depth to ensure termination. If this goes over a certain threshold, we abort compilation – in such cases, we can not say whether or not the predicate holds for certain. Stupid halting problem; such a drag.

Implementations

impl<'tcx> Obligation<'tcx, Predicate<'tcx>>

pub fn flip_polarity( &self, tcx: TyCtxt<'tcx>, ) -> Option<Obligation<'tcx, Predicate<'tcx>>>

Flips the polarity of the inner predicate.

Given T: Trait predicate it returns T: !Trait and given T: !Trait returns T: Trait.

impl<'tcx> Obligation<'tcx, Binder<TyCtxt<'tcx>, TraitPredicate<TyCtxt<'tcx>>>>

pub fn derived_cause( &self, variant: impl FnOnce(DerivedCause<'tcx>) -> ObligationCauseCode<'tcx>, ) -> ObligationCause<'tcx>

impl<'tcx, O> Obligation<'tcx, O>

pub fn new( tcx: TyCtxt<'tcx>, cause: ObligationCause<'tcx>, param_env: ParamEnv<'tcx>, predicate: impl Upcast<TyCtxt<'tcx>, O>, ) -> Obligation<'tcx, O>

pub fn set_depth_from_parent(&mut self, parent_depth: usize)

We often create nested obligations without setting the correct depth.

To deal with this evaluate and fulfill explicitly update the depth of nested obligations using this function.

pub fn with_depth( tcx: TyCtxt<'tcx>, cause: ObligationCause<'tcx>, recursion_depth: usize, param_env: ParamEnv<'tcx>, predicate: impl Upcast<TyCtxt<'tcx>, O>, ) -> Obligation<'tcx, O>

pub fn misc( tcx: TyCtxt<'tcx>, span: Span, body_id: LocalDefId, param_env: ParamEnv<'tcx>, trait_ref: impl Upcast<TyCtxt<'tcx>, O>, ) -> Obligation<'tcx, O>

pub fn with<P>( &self, tcx: TyCtxt<'tcx>, value: impl Upcast<TyCtxt<'tcx>, P>, ) -> Obligation<'tcx, P>

impl<'tcx> Obligation<'tcx, Binder<TyCtxt<'tcx>, TraitPredicate<TyCtxt<'tcx>>>>

pub fn polarity(&self) -> PredicatePolarity

pub fn self_ty(&self) -> Binder<TyCtxt<'tcx>, Ty<'tcx>>