Trait rustc_trait_selection::traits::TraitEngine

pub trait TraitEngine<'tcx>: 'tcx {
    fn register_predicate_obligation(
        &mut self,
        infcx: &InferCtxt<'tcx>,
        obligation: Obligation<'tcx, Predicate<'tcx>>
    );
    fn select_all_or_error(
        &mut self,
        infcx: &InferCtxt<'tcx>
    ) -> Vec<FulfillmentError<'tcx>, Global>;
    fn select_where_possible(
        &mut self,
        infcx: &InferCtxt<'tcx>
    ) -> Vec<FulfillmentError<'tcx>, Global>;
    fn pending_obligations(
        &self
    ) -> Vec<Obligation<'tcx, Predicate<'tcx>>, Global>;
    fn relationships(
        &mut self
    ) -> &mut HashMap<TyVid, FoundRelationships, BuildHasherDefault<FxHasher>>;

    fn register_bound(
        &mut self,
        infcx: &InferCtxt<'tcx>,
        param_env: ParamEnv<'tcx>,
        ty: Ty<'tcx>,
        def_id: DefId,
        cause: ObligationCause<'tcx>
    ) { ... }
}

Required Methods

fn register_predicate_obligation(
&mut self,
infcx: &InferCtxt<'tcx>,
obligation: Obligation<'tcx, Predicate<'tcx>>
)

fn select_all_or_error(
&mut self,
infcx: &InferCtxt<'tcx>
) -> Vec<FulfillmentError<'tcx>, Global>

fn select_where_possible(
&mut self,
infcx: &InferCtxt<'tcx>
) -> Vec<FulfillmentError<'tcx>, Global>

fn pending_obligations(&self) -> Vec<Obligation<'tcx, Predicate<'tcx>>, Global>

fn relationships(
&mut self
) -> &mut HashMap<TyVid, FoundRelationships, BuildHasherDefault<FxHasher>>

Provided Methods

fn register_bound(
&mut self,
infcx: &InferCtxt<'tcx>,
param_env: ParamEnv<'tcx>,
ty: Ty<'tcx>,
def_id: DefId,
cause: ObligationCause<'tcx>
)

Requires that ty must implement the trait with def_id in the given environment. This trait must not have any type parameters (except for Self).