rustc_trait_selection::traits
Type Alias SelectionResult

pub type SelectionResult<'tcx, T> = Result<Option<T>, SelectionError<'tcx>>;

When performing resolution, it is typically the case that there can be one of three outcomes:

Ok(Some(r)): success occurred with result r

Ok(None): could not definitely determine anything, usually due to inconclusive type inference.

Err(e): error e occurred

Aliased Type

enum SelectionResult<'tcx, T> {
    Ok(Option<T>),
    Err(SelectionError<'tcx>),
}

Variants

Ok(Option<T>)

Contains the success value

Err(SelectionError<'tcx>)

Contains the error value

Layout

Note: Unable to compute type layout, possibly due to this type having generic parameters. Layout can only be computed for concrete, fully-instantiated types.