Module rustc_trait_selection::solve

Modules

delegate 🔒

fulfill 🔒

inspect

normalize 🔒

select 🔒

Structs

CacheData

EvalCtxt

ExternalConstraintsData

Additional constraints returned on success.

FulfillmentCtxt

A trait engine using the new trait solver.

Goal

A goal is a statement, i.e. predicate, we want to prove given some assumptions, i.e. param_env.

NestedNormalizationGoals

NoSolution

PredefinedOpaquesData

Opaques that are defined in the inference context before a query is called.

QueryInput

Response

Enums

BuiltinImplSource

CandidateSource

Possible ways the given goal can be proven.

Certainty

GenerateProofTree

GoalSource

Why a specific goal has to be proven.

MaybeCause

Why we failed to evaluate a goal.

NextSolverError

Reveal

Depending on the stage of compilation, we want projection to be more or less conservative.

SolverMode

Traits

InferCtxtSelectExt

SolverDelegateEvalExt

Functions

deeply_normalize

Deeply normalize all aliases in value. This does not handle inference and expects its input to be already fully resolved.

deeply_normalize_with_skipped_universes

Deeply normalize all aliases in value. This does not handle inference and expects its input to be already fully resolved.

Type Aliases

CanonicalInput

CanonicalResponse

QueryResult

The result of evaluating a canonical query.