Module rustc_trait_selection::traits::coherence

See Rustc Dev Guide chapters on trait-resolution and trait-specialization for more info on how this works.

Structs

OrphanChecker 🔒

OverlapResult

Enums

Conflict

InCrate 🔒

Whether we do the orphan check relative to this crate or to some remote crate.

OrphanCheckEarlyExit 🔒

OrphanCheckErr

TrackAmbiguityCauses 🔒

Functions

add_placeholder_note

equate_impl_headers 🔒

impl_intersection_has_impossible_obligation 🔒

Check if both impls can be satisfied by a common type by considering whether any of either impl’s obligations is not known to hold.

impl_intersection_has_negative_obligation 🔒

Check if both impls can be satisfied by a common type by considering whether any of first impl’s obligations is known not to hold via a negative predicate.

negative_impl_exists 🔒

Try to prove that a negative impl exist for the obligation or its supertraits.

orphan_check

Checks the coherence orphan rules. impl_def_id should be the DefId of a trait impl. To pass, either the trait must be local, or else two conditions must be satisfied:

orphan_check_trait_ref 🔒

Checks whether a trait-ref is potentially implementable by a crate.

overlap 🔒

Can both impl a and impl b be satisfied by a common type (including where-clauses)? If so, returns an ImplHeader that unifies the two impls.

overlapping_impls

If there are types that satisfy both impls, returns Some with a suitably-freshened ImplHeader with those types substituted. Otherwise, returns None.

prove_negated_obligation 🔒

trait_ref_is_knowable

Returns whether all impls which would apply to the trait_ref e.g. Ty: Trait<Arg> are already known in the local crate.

trait_ref_is_local_or_fundamental

with_fresh_ty_vars 🔒