Struct rustc_trait_selection::traits::fulfill::FulfillmentContext

pub struct FulfillmentContext<'tcx> {
    predicates: ObligationForest<PendingPredicateObligation<'tcx>>,
    usable_in_snapshot: usize,
}

The fulfillment context is used to drive trait resolution. It consists of a list of obligations that must be (eventually) satisfied. The job is to track which are satisfied, which yielded errors, and which are still pending. At any point, users can call select_where_possible, and the fulfillment context will try to do selection, retaining only those obligations that remain ambiguous. This may be helpful in pushing type inference along. Once all type inference constraints have been generated, the method select_all_or_error can be used to report any remaining ambiguous cases as errors.

Fields
predicates: ObligationForest<PendingPredicateObligation<'tcx>>

A list of all obligations that have been registered with this fulfillment context.

usable_in_snapshot: usize

The snapshot in which this context was created. Using the context outside of this snapshot leads to subtle bugs if the snapshot gets rolled back. Because of this we explicitly check that we only use the context in exactly this snapshot.

Implementations

impl<'tcx> FulfillmentContext<'tcx>

pub(super) fn new(infcx: &InferCtxt<'tcx>) -> FulfillmentContext<'tcx>

Creates a new fulfillment context.

fn select( &mut self, selcx: SelectionContext<'_, 'tcx> ) -> Vec<FulfillmentError<'tcx>>

Attempts to select obligations using selcx.