rustc_trait_selection::traits::query
Module dropck_outlives

Functions

compute_dropck_outlives_inner

dtorck_constraint_for_ty_inner

Returns a set of constraints that needs to be satisfied in order for ty to be valid for destruction.

trivial_dropck_outlives

This returns true if the type ty is “trivial” for dropck-outlives – that is, if it doesn’t require any types to outlive. This is similar but not quite the same as the needs_drop test in the compiler already – that is, for every type T for which this function return true, needs-drop would return false. But the reverse does not hold: in particular, needs_drop returns false for PhantomData, but it is not trivial for dropck-outlives.