rustc_trait_selection::traits::query
Module type_op

Modules

ascribe_user_type

custom

implied_outlives_bounds

normalize

outlives

prove_predicate

Structs

AscribeUserType

DropckOutlives

Eq

ImpliedOutlivesBounds

Normalize

ProvePredicate

Subtype

TypeOpOutput

The output from performing a type op

Traits

QueryTypeOp

“Query type ops” are type ops that are implemented using a canonical query. The Self type here contains the kernel of information needed to do the operation – TypeOp is actually implemented for ParamEnvAnd<Self>, since we always need to bring along a parameter environment as well. For query type-ops, we will first canonicalize the key and then invoke the query on the tcx, which produces the resulting query region constraints.

TypeOp

“Type ops” are used in NLL to perform some particular action and extract out the resulting region constraints (or an error if it cannot be completed).