Trait rustc_trait_selection::traits::query::type_op::TypeOp

pub trait TypeOp<'tcx>: Sized + Debug {
    type Output: Debug;
    type ErrorInfo;

    // Required method
    fn fully_perform(
        self,
        infcx: &InferCtxt<'tcx>,
        span: Span
    ) -> Result<TypeOpOutput<'tcx, Self>, ErrorGuaranteed>;
}

“Type ops” are used in NLL to perform some particular action and extract out the resulting region constraints (or an error if it cannot be completed).

Required Associated Types

type Output: Debug

type ErrorInfo

Required Methods

fn fully_perform( self, infcx: &InferCtxt<'tcx>, span: Span ) -> Result<TypeOpOutput<'tcx, Self>, ErrorGuaranteed>

Processes the operation and all resulting obligations, returning the final result along with any region constraints (they will be given over to the NLL region solver).

Object Safety

This trait is not object safe.