Module rustc_trait_selection::traits::specialize::specialization_graph

Structs

Ancestors

Children

Children of a given impl, grouped into blanket/non-blanket varieties as is done in TraitDef.

FutureCompatOverlapError

Graph

A per-trait graph of impls in specialization order. At the moment, this graph forms a tree rooted with the trait itself, with all other nodes representing impls, and parent-child relationships representing specializations.

LeafDef

Information about the most specialized definition of an associated item.

Enums

FutureCompatOverlapErrorKind

Inserted 🔒

The result of attempting to insert an impl into a group of children.

Node

A node in the specialization graph is either an impl or a trait definition; either can serve as a source of item definitions. There is always exactly one trait definition node: the root.

OverlapMode

What kind of overlap check are we doing – this exists just for testing and feature-gating purposes.

PotentialSiblings 🔒

Traits

ChildrenExt 🔒

GraphExt

Functions

ancestors

Walk up the specialization ancestors of a given impl, starting with that impl itself.

assoc_def 🔒

Locate the definition of an associated type in the specialization hierarchy, starting from the given impl.

filtered_children 🔒

iter_children 🔒