Trait rustc_trait_selection::traits::TraitEngine

pub trait TraitEngine<'tcx>: 'tcx {
    // Required methods
    fn register_predicate_obligation(
        &mut self,
        infcx: &InferCtxt<'tcx>,
        obligation: Obligation<'tcx, Predicate<'tcx>>
    );
    fn select_where_possible(
        &mut self,
        infcx: &InferCtxt<'tcx>
    ) -> Vec<FulfillmentError<'tcx>>;
    fn collect_remaining_errors(
        &mut self,
        infcx: &InferCtxt<'tcx>
    ) -> Vec<FulfillmentError<'tcx>>;
    fn pending_obligations(&self) -> Vec<Obligation<'tcx, Predicate<'tcx>>>;
    fn drain_unstalled_obligations(
        &mut self,
        infcx: &InferCtxt<'tcx>
    ) -> Vec<Obligation<'tcx, Predicate<'tcx>>>;

    // Provided method
    fn register_bound(
        &mut self,
        infcx: &InferCtxt<'tcx>,
        param_env: ParamEnv<'tcx>,
        ty: Ty<'tcx>,
        def_id: DefId,
        cause: ObligationCause<'tcx>
    ) { ... }
}

Required Methods

fn register_predicate_obligation( &mut self, infcx: &InferCtxt<'tcx>, obligation: Obligation<'tcx, Predicate<'tcx>> )

fn select_where_possible( &mut self, infcx: &InferCtxt<'tcx> ) -> Vec<FulfillmentError<'tcx>>

fn collect_remaining_errors( &mut self, infcx: &InferCtxt<'tcx> ) -> Vec<FulfillmentError<'tcx>>

fn pending_obligations(&self) -> Vec<Obligation<'tcx, Predicate<'tcx>>>

fn drain_unstalled_obligations( &mut self, infcx: &InferCtxt<'tcx> ) -> Vec<Obligation<'tcx, Predicate<'tcx>>>

Among all pending obligations, collect those are stalled on a inference variable which has changed since the last call to select_where_possible. Those obligations are marked as successful and returned.

Provided Methods

fn register_bound( &mut self, infcx: &InferCtxt<'tcx>, param_env: ParamEnv<'tcx>, ty: Ty<'tcx>, def_id: DefId, cause: ObligationCause<'tcx> )

Requires that ty must implement the trait with def_id in the given environment. This trait must not have any type parameters (except for Self).