Module rustc_trait_selection::traits::util

Structs

BoundVarReplacer

Elaborator

“Elaboration” is the process of identifying all the predicates that are implied by a source predicate. Currently, this basically means walking the “supertraits” and other similar assumptions. For example, if we know that T: Ord, the elaborator would deduce that T: PartialOrd holds as well. Similarly, if we have trait Foo: 'static, and we know that T: Foo, then we know that T: 'static.

FilterToTraits

A filter around an iterator of predicates that makes it yield up just trait references.

PlaceholderReplacer

The inverse of BoundVarReplacer: replaces placeholders with the bound vars from which they came.

PredicateSet

TraitAliasExpander

“Trait alias expansion” is the process of expanding a sequence of trait references into another sequence by transitively following all trait aliases. e.g. If you have bounds like Foo + Send, a trait alias trait Foo = Bar + Sync;, and another trait alias trait Bar = Read + Write, then the bounds would expand to Read + Write + Sync + Send. Expansion is done via a DFS (depth-first search), and the visited field is used to avoid cycles.

TraitAliasExpansionInfo

Stores information about the expansion of a trait via a path of zero or more trait aliases.

Enums

TupleArgumentsFlag

Traits

Elaboratable

Describes how to elaborate an obligation into a sub-obligation.

Functions

anonymize_predicate

async_iterator_trait_ref_and_outputs

closure_trait_ref_and_return_type

coroutine_trait_ref_and_outputs

elaborate

expand_trait_aliases

future_trait_ref_and_outputs

impl_item_is_final

impl_subject_and_oblig

Instantiate all bound parameters of the impl subject with the given args, returning the resulting subject and all obligations that arise. The obligations are closed under normalization.

iterator_trait_ref_and_outputs

supertrait_def_ids

Computes the def-ids of the transitive supertraits of trait_def_id. This (intentionally) does not compute the full elaborated super-predicates but just the set of def-ids. It is used to identify which traits may define a given associated type to help avoid cycle errors, and to make size estimates for vtable layout computation.

supertraits

transitive_bounds

transitive_bounds_that_define_assoc_item

A specialized variant of elaborate that only elaborates trait references that may define the given associated item with the name assoc_name. It uses the explicit_supertraits_containing_assoc_item query to avoid enumerating super-predicates that aren’t related to assoc_item. This is used when resolving types like Self::Item or T::Item and helps to avoid cycle errors (see e.g. #35237).

upcast_choices

Casts a trait reference into a reference to one of its super traits; returns None if target_trait_def_id is not a supertrait.

with_replaced_escaping_bound_vars

Executes f on value after replacing all escaping bound variables with placeholders and then replaces these placeholders with the original bound variables in the result.